pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[project]
name = "creative-automation-pipeline"
version = "0.1.0"
description = "AI-powered creative automation for scalable social ad campaigns"
requires-python = ">=3.10,<3.14"
authors = [
    {name = "Creative Automation Team"}
]
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pillow==11.0.0",
    "pyyaml==6.0.1",
    "openai==1.3.0",
    "python-multipart==0.0.6",
    "python-dotenv==1.0.0",
    "pydantic==2.9.0",
    "pydantic-settings==2.4.0",
    "requests==2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "responses==0.26.0",
    "black==23.11.0",
    "ruff==0.1.7",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra"
